backend/requirements.txt
Django==2.2.19
pytz==2023.3
sqlparse==0.4.3
Pillow==8.3.1
psycopg2-binary==2.8.6
python-dotenv
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==2.4.0
reportlab
webcolors==1.11.1
gunicorn==20.0.4